requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
auth0-python==3.24.1
Authlib==1.4.1
bcrypt==4.2.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==3.3.2
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
Flask==3.1.0
Flask-Authlib==1.6.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jwt==1.3.1
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysql-connector-python==9.2.0
openai==1.61.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.5.0
PyMySQL==1.1.1
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.32.3
rich==13.9.4
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
tqdm==4.67.1
types-cryptography==3.3.23.2
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3